requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
pytesseract
numpy
Pillow
pdf2image
google-generativeai
